+++ tests/sync/test_chat.py
@@ -3,6 +3,14 @@ import unittest
 from utils import get_api_key
 
 import cohere
+from cohere.responses.chat import (
+    StreamCitationGeneration,
+    StreamEnd,
+    StreamQueryGeneration,
+    StreamSearchResults,
+    StreamStart,
+    StreamTextGeneration,
+)
 
 API_KEY = get_api_key()
 co = cohere.Client(API_KEY)
@@ -12,7 +20,6 @@ class TestChat(unittest.TestCase):
     def test_simple_success(self):
         prediction = co.chat("Yo what up?", max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
         self.assertTrue(prediction.meta)
         self.assertTrue(prediction.meta["api_version"])
         self.assertTrue(prediction.meta["api_version"]["version"])
@@ -23,12 +30,10 @@ class TestChat(unittest.TestCase):
         for _ in range(num_replies):
             prediction = prediction.respond("oh that's cool", max_tokens=5)
             self.assertIsInstance(prediction.text, str)
-            self.assertIsInstance(prediction.conversation_id, str)
 
     def test_valid_model(self):
-        prediction = co.chat("Yo what up?", model="medium", max_tokens=5)
+        prediction = co.chat("Yo what up?", model="command", max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
 
     def test_invalid_model(self):
         with self.assertRaises(cohere.CohereError):
@@ -37,28 +42,24 @@ class TestChat(unittest.TestCase):
     def test_return_chatlog(self):
         prediction = co.chat("Yo what up?", return_chatlog=True, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
         self.assertIsNotNone(prediction.chatlog)
         self.assertGreaterEqual(len(prediction.chatlog), len(prediction.text))
 
     def test_return_chatlog_false(self):
         prediction = co.chat("Yo what up?", return_chatlog=False, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
 
         assert prediction.chatlog is None
 
     def test_return_prompt(self):
         prediction = co.chat("Yo what up?", return_prompt=True, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
         self.assertIsNotNone(prediction.prompt)
         self.assertGreaterEqual(len(prediction.prompt), len(prediction.text))
 
     def test_return_prompt_false(self):
         prediction = co.chat("Yo what up?", return_prompt=False, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
         assert prediction.prompt is None
 
     def test_preamble_override(self):
@@ -67,7 +68,6 @@ class TestChat(unittest.TestCase):
             "Yo what up?", preamble_override=preamble, return_prompt=True, return_preamble=True, max_tokens=5
         )
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
         self.assertIn(preamble, prediction.prompt)
         self.assertEqual(preamble, prediction.preamble)
 
@@ -82,7 +82,6 @@ class TestChat(unittest.TestCase):
         for temperature in temperatures:
             prediction = co.chat("Yo what up?", temperature=temperature, max_tokens=5)
             self.assertIsInstance(prediction.text, str)
-            self.assertIsInstance(prediction.conversation_id, str)
 
     def test_stream(self):
         prediction = co.chat(
@@ -94,24 +93,25 @@ class TestChat(unittest.TestCase):
         self.assertIsInstance(prediction, cohere.responses.chat.StreamingChat)
         self.assertIsInstance(prediction.texts, list)
         self.assertEqual(len(prediction.texts), 0)
-        self.assertIsNone(prediction.conversation_id)
         self.assertIsNone(prediction.response_id)
         self.assertIsNone(prediction.finish_reason)
 
         expected_index = 0
         expected_text = ""
         for token in prediction:
-            self.assertIsInstance(token.text, str)
-            self.assertGreater(len(token.text), 0)
-
+            if isinstance(token, cohere.responses.chat.StreamStart):
+                self.assertIsNotNone(token.generation_id)
+                self.assertFalse(token.is_finished)
+            elif isinstance(token, cohere.responses.chat.StreamTextGeneration):
+                self.assertIsInstance(token.text, str)
+                self.assertGreater(len(token.text), 0)
+                expected_text += token.text
+                self.assertFalse(token.is_finished)
             self.assertIsInstance(token.index, int)
             self.assertEqual(token.index, expected_index)
-
-            expected_text += token.text
             expected_index += 1
 
         self.assertEqual(prediction.texts, [expected_text])
-        self.assertIsNotNone(prediction.conversation_id)
         self.assertIsNotNone(prediction.response_id)
         self.assertIsNotNone(prediction.finish_reason)
 
@@ -127,7 +127,6 @@ class TestChat(unittest.TestCase):
     def test_return_preamble(self):
         prediction = co.chat("Yo what up?", return_preamble=True, return_prompt=True, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
         self.assertIsNotNone(prediction.preamble)
         self.assertIsNotNone(prediction.prompt)
         self.assertIn(prediction.preamble, prediction.prompt)
@@ -135,7 +134,6 @@ class TestChat(unittest.TestCase):
     def test_return_preamble_false(self):
         prediction = co.chat("Yo what up?", return_preamble=False, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
 
         assert prediction.preamble is None
 
@@ -151,7 +149,6 @@ class TestChat(unittest.TestCase):
             max_tokens=5,
         )
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
         self.assertIsNotNone(prediction.chatlog)
         self.assertIn("User: Hey!", prediction.prompt)
         self.assertIn("Chatbot: Hey! How can I help you?", prediction.prompt)
@@ -181,7 +178,6 @@ class TestChat(unittest.TestCase):
     def test_p(self):
         prediction = co.chat("Yo what up?", p=0.9, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
 
     def test_invalid_p(self):
         with self.assertRaises(cohere.error.CohereError):
@@ -190,7 +186,6 @@ class TestChat(unittest.TestCase):
     def test_k(self):
         prediction = co.chat("Yo what up?", k=5, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
 
     def test_invalid_k(self):
         with self.assertRaises(cohere.error.CohereError):
@@ -199,7 +194,6 @@ class TestChat(unittest.TestCase):
     def test_logit_bias(self):
         prediction = co.chat("Yo what up?", logit_bias={42: 10}, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsInstance(prediction.conversation_id, str)
 
     def test_invalid_logit_bias(self):
         invalid = [
@@ -211,3 +205,156 @@ class TestChat(unittest.TestCase):
         for logit_bias in invalid:
             with self.assertRaises(cohere.error.CohereError):
                 _ = co.chat("Yo what up?", logit_bias=logit_bias, max_tokens=5)
+
+    def test_search_queries_only_true(self):
+        prediction = co.chat(
+            "What is the height of Mount Everest? What is the depth of the Mariana Trench? What is the climate like in Nepal?",
+            search_queries_only=True,
+        )
+        self.assertTrue(prediction.is_search_required)
+        self.assertIsInstance(prediction.search_queries, list)
+        self.assertGreater(len(prediction.search_queries), 0)
+        self.assertIsInstance(prediction.search_queries[0]["text"], str)
+        self.assertIsInstance(prediction.search_queries[0]["generation_id"], str)
+
+    def test_search_queries_only_false(self):
+        prediction = co.chat("hello", search_queries_only=True)
+        self.assertFalse(prediction.is_search_required)
+        self.assertIsInstance(prediction.search_queries, list)
+        self.assertEqual(len(prediction.search_queries), 0)
+
+    def test_with_documents(self):
+        prediction = co.chat(
+            "How deep in the Mariana Trench",
+            temperature=0,
+            documents=[
+                {
+                    "id": "national_geographic_everest",
+                    "title": "Height of Mount Everest",
+                    "snippet": "The height of Mount Everest is 29,035 feet",
+                    "url": "https://education.nationalgeographic.org/resource/mount-everest/",
+                },
+                {
+                    "id": "national_geographic_mariana",
+                    "title": "Depth of the Mariana Trench",
+                    "snippet": "The depth of the Mariana Trench is 36,070 feet",
+                    "url": "https://www.nationalgeographic.org/activity/mariana-trench-deepest-place-earth",
+                },
+            ],
+        )
+        self.assertIsInstance(prediction.text, str)
+        self.assertIsInstance(prediction.citations, list)
+        self.assertGreater(len(prediction.citations), 0)
+        self.assertIsInstance(prediction.citations[0]["start"], int)
+        self.assertIsInstance(prediction.citations[0]["end"], int)
+        self.assertIsInstance(prediction.citations[0]["text"], str)
+        self.assertIsInstance(prediction.citations[0]["document_ids"], list)
+        self.assertGreater(len(prediction.citations[0]["document_ids"]), 0)
+        self.assertIsInstance(prediction.documents, list)
+        self.assertGreater(len(prediction.documents), 0)
+
+    def test_with_connectors(self):
+        prediction = co.chat("How deep in the Mariana Trench", temperature=0, connectors=[{"id": "web-search"}])
+        self.assertIsInstance(prediction.text, str)
+        self.assertIsInstance(prediction.citations, list)
+        self.assertGreater(len(prediction.citations), 0)
+        self.assertIsInstance(prediction.citations[0]["start"], int)
+        self.assertIsInstance(prediction.citations[0]["end"], int)
+        self.assertIsInstance(prediction.citations[0]["text"], str)
+        self.assertIsInstance(prediction.citations[0]["document_ids"], list)
+        self.assertGreater(len(prediction.citations[0]["document_ids"]), 0)
+        self.assertIsInstance(prediction.documents, list)
+        self.assertGreater(len(prediction.documents), 0)
+        self.assertIsInstance(prediction.search_results, list)
+        self.assertGreater(len(prediction.search_results), 0)
+
+    def test_with_citation_quality(self):
+        prediction = co.chat(
+            "How deep in the Mariana Trench",
+            citation_quality="accurate",
+            temperature=0,
+            documents=[
+                {
+                    "id": "national_geographic_mariana",
+                    "title": "Depth of the Mariana Trench",
+                    "snippet": "The depth of the Mariana Trench is 36,070 feet",
+                    "url": "https://www.nationalgeographic.org/activity/mariana-trench-deepest-place-earth",
+                },
+            ],
+        )
+        self.assertIsInstance(prediction.text, str)
+        self.assertIsInstance(prediction.citations, list)
+        self.assertGreater(len(prediction.citations), 0)
+        self.assertIsInstance(prediction.citations[0]["start"], int)
+        self.assertIsInstance(prediction.citations[0]["end"], int)
+        self.assertIsInstance(prediction.citations[0]["text"], str)
+        self.assertIsInstance(prediction.citations[0]["document_ids"], list)
+        self.assertGreater(len(prediction.citations[0]["document_ids"]), 0)
+        self.assertIsInstance(prediction.documents, list)
+        self.assertGreater(len(prediction.documents), 0)
+
+    def test_stream_with_connectors(self):
+        prediction = co.chat(
+            "How deep in the Mariana Trench", temperature=0, stream=True, connectors=[{"id": "web-search"}]
+        )
+
+        self.assertIsInstance(prediction, cohere.responses.chat.StreamingChat)
+        self.assertIsInstance(prediction.texts, list)
+        self.assertEqual(len(prediction.texts), 0)
+        self.assertIsNone(prediction.response_id)
+        self.assertIsNone(prediction.finish_reason)
+
+        expected_index = 0
+        expected_text = ""
+
+        count_stream_start = 0
+        count_text_generation = 0
+        count_query_generation = 0
+        count_citation_generation = 0
+        count_search_results = 0
+        count_stream_end = 0
+        for token in prediction:
+            if isinstance(token, StreamStart):
+                count_stream_start += 1
+                self.assertIsNotNone(token.generation_id)
+                self.assertFalse(token.is_finished)
+                self.assertEqual(token.event_type, "stream-start")
+            elif isinstance(token, StreamQueryGeneration):
+                count_query_generation += 1
+                self.assertIsNotNone(token.search_queries)
+                self.assertEqual(token.event_type, "search-queries-generation")
+            elif isinstance(token, StreamSearchResults):
+                count_search_results += 1
+                self.assertIsNotNone(token.documents)
+                self.assertIsNotNone(token.search_results)
+                self.assertEqual(token.event_type, "search-results")
+            elif isinstance(token, StreamCitationGeneration):
+                count_citation_generation += 1
+                self.assertIsNotNone(token.citations)
+                self.assertEqual(token.event_type, "citation-generation")
+            elif isinstance(token, StreamTextGeneration):
+                count_text_generation += 1
+                self.assertIsInstance(token.text, str)
+                self.assertGreater(len(token.text), 0)
+                expected_text += token.text
+                self.assertFalse(token.is_finished)
+                self.assertEqual(token.event_type, "text-generation")
+            elif isinstance(token, StreamEnd):
+                count_stream_end += 1
+                self.assertTrue(token.is_finished)
+                self.assertEqual(token.event_type, "stream-end")
+                self.assertEqual(token.finish_reason, "COMPLETE")
+            self.assertIsInstance(token.index, int)
+            self.assertEqual(token.index, expected_index)
+            expected_index += 1
+
+        self.assertEqual(count_stream_start, 1)
+        self.assertEqual(count_search_results, 1)
+        self.assertEqual(count_stream_end, 1)
+        self.assertGreaterEqual(count_citation_generation, 1)
+        self.assertGreaterEqual(count_query_generation, 1)
+        self.assertGreaterEqual(count_text_generation, 1)
+
+        self.assertEqual(prediction.texts, [expected_text])
+        self.assertIsNotNone(prediction.response_id)
+        self.assertIsNotNone(prediction.finish_reason)